Removed templates Only int Matrix
